Full-Orchestrator/requirements.txt
# Full-Orchestrator: generate infra + app from requirements
crewai>=0.80.0
crewai-tools>=0.14.0
requests>=2.28.0
boto3>=1.26.0
python-dotenv>=1.0.0
